refactoring: fixed imports of useAuth in projects.
