+++ src/modules/auth/index.ts
@@ -1,4 +1,2 @@
-export { default as useAuth } from './hooks/api/useAuth';
-
 export { default as LogOutButton } from './ui/containers/LogOutButton';
 export { default as LogInForm } from './ui/containers/LogInFormContainer';
