+++ src/shared/api/index.ts
@@ -5,3 +5,4 @@ export * as METHODS from './config/methods';
 export * as TAGS from './config/invalidation-tags';
 
 export { default as baseAPI } from './model/baseApi';
+export { default as useAuth } from './model/useAuth';
